Fix circular dependency warning in table #LMR-1007
