--- src/app/view/perspectives/table2/header/column-group/compound-column/table-compound-column.component.html
@@ -1,24 +1,0 @@
-<div mwlResizable
-     [enableGhostResize]="true"
-     [ghostElementPositioning]="'absolute'"
-     [resizeEdges]="column | resizeEdges:table:cursor"
-     [resizeSnapGrid]="{left: 10, right: 10}"
-     [style.width.px]="column | columnWidth"
-     [validateResize]="validateResize.bind(this)"
-     (resizeEnd)="onResizeEnd($event)">
-  <div class="d-flex flex-column">
-    <table-single-column [table]="table"
-                         [cursor]="cursor"
-                         [column]="column.parent"
-                         [leaf]="!column.children.length"
-                         [style.height.px]="column | columnHeight:table:cursor"
-                         class="d-block">
-    </table-single-column>
-    <table-column-group *ngIf="column.children.length"
-                        [table]="table"
-                        [cursor]="cursor"
-                        [columns]="column.children"
-                        class="d-block">
-    </table-column-group>
-  </div>
-</div>
