--- src/app/view/perspectives/table2/header/column-group/compound-column/table-compound-column.component.scss
@@ -1,3 +1,0 @@
-mwlResizable {
-  box-sizing: border-box; // required for the enableGhostResize option to work
-}
